F ut

+++ be/test/exec/sink/vtablet_writer_v2_test.cpp
@@ -148,8 +148,8 @@ TEST_F(TestVTabletWriterV2, load_stream_reply_profile_should_be_collected) {
     RuntimeProfile remote_profile("LoadStream");
     auto* delta_writer_profile = remote_profile.create_child("DeltaWriterV2 10001", true, true);
     auto* memtable_writer_profile = remote_profile.create_child("MemTableWriter 10001", true, true);
-    COUNTER_SET(ADD_TIMER(delta_writer_profile, "WriteMemTableTime"), 1);
-    COUNTER_SET(ADD_TIMER(memtable_writer_profile, "MemTableSortTime"), 1);
+    COUNTER_SET(ADD_TIMER(delta_writer_profile, "WriteMemTableTime"), static_cast<int64_t>(1));
+    COUNTER_SET(ADD_TIMER(memtable_writer_profile, "MemTableSortTime"), static_cast<int64_t>(1));
     TRuntimeProfileTree remote_profile_tree;
     remote_profile.to_thrift(&remote_profile_tree);
